split dashboard from cluster
